requirements.txt
Django==1.11.2
djangorestframework==3.6.3
Pygments==2.2.0
pytz==2017.2
